2026-01-21 | (Backend): The use of int is replaced with long (Int64) for calculating the Collatz sequence, preventing overflow in operations such as 3n + 1. The theoretical maximum range of the input value increases from 2,100,000,000 to 3,074,457,345,618,258,602, limited by the capacity of the long type. Explicit overflow control is incorporated using checked, allowing for the detection of out-of-range values. The input parameter is received as a string and internally converted to long, ensuring accuracy for large values. The API response is returned as string[], preventing loss of precision in JavaScript clients. (Frontend): The umber type is no longer used to handle large values ​​due to JavaScript's 2⁵³ − 1 limit. Range validation is performed using BigInt, ensuring exact comparisons for large numbers. The user-entered value is sent to the backend as a string, preserving accuracy during HTTP communication. The use of numeric literals outside JavaScript's safe range is avoided, eliminating compilation errors and inconsistent results.

+++ backend/API/Controllers/CollatzConjectureController.cs
@@ -18,9 +18,9 @@ public class CollatzConjectureController : BaseApiController
     [HttpGet("service")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    public Task <ActionResult<List<int>>> Get(int value)
+    public Task <ActionResult<List<long>>> Get(long value)
     {
-        List<int> list = new List<int>();
+        List<long> list = new List<long>();
         list.Add(value);
         try
         {
@@ -34,11 +34,11 @@ public class CollatzConjectureController : BaseApiController
                 list.Add(value);
             }
 
-            return Task.FromResult<ActionResult<List<int>>>(list);
+            return Task.FromResult<ActionResult<List<long>>>(list);
         }
         catch (Exception ex)
         {
-            return Task.FromResult<ActionResult<List<int>>>(StatusCode(500, new { Message = "There was an issue retrieving the continents. Please try again later.", Details = ex.Message }));
+            return Task.FromResult<ActionResult<List<long>>>(StatusCode(500, new { Message = "There was an issue retrieving the continents. Please try again later.", Details = ex.Message }));
         }
     }
